Move XorTransport into transports package
