+++ kasa/device_factory.py
@@ -8,6 +8,7 @@ from typing import Any
 
 from kasa.transports.aestransport import AesTransport
 from kasa.transports.klaptransport import KlapTransport, KlapTransportV2
+from kasa.transports.xortransport import XorTransport
 
 from .device import Device
 from .device_type import DeviceType
@@ -32,7 +33,6 @@ from .protocol import (
 )
 from .smart import SmartDevice
 from .smartprotocol import SmartProtocol
-from .xortransport import XorTransport
 
 _LOGGER = logging.getLogger(__name__)
